backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
pydantic-settings
python-dotenv
httpx
python-jose[cryptography]
passlib[bcrypt]
python-multipart
scikit-learn
numpy
websockets
aiofiles
requests
tinydb
python-dateutil
